统一使用 getGeneratedKeys 回填自增主键，移除两段式 last_insert_id
- core: 新增 SqlBeanUtil.setAutoIncrId(clazz, keys, beans...) 按索引 1:1 回填 @SqlId(type=AUTO) 字段，取代原先依赖连续分配的假设
- core: 删除 SqlBeanProvider.lastInsertIdSql() 两段式实现
- core: SqlBeanException 增加 (String, Throwable) 构造器以保留原始异常
- BaseSqlBeanServiceImpl: 移除抽象 getAutoIncrId()，setAutoIncrId 改为接收真实生成键列表
- Spring-JDBC: 改用 RETURN_GENERATED_KEYS + GeneratedKeyHolder 取键回填（单条/列表/Insert 三处）
- MyBatis(Spring): 通过当前 SqlSession 连接 PreparedStatement+RETURN_GENERATED_KEYS 逐 bean 取键回填，保留 @DbSwitch 路由与事务
- MyBatis(Solon): 通过事务连接 PreparedStatement+RETURN_GENERATED_KEYS 逐 bean 取键回填
- Android: SQLiteTemplate.insert 返回 rowid(long)，insert 改为事务内逐 bean 循环回填（支持批量）
- 删除 MybatisSqlBeanDao.lastInsertId / MybatisSqlBeanProvider.lastInsertId

+++ flexsql-android/src/main/java/cn/vonce/sql/android/helper/SQLiteTemplate.java
@@ -81,14 +81,14 @@ public class SQLiteTemplate {
     }
 
     /**
-     * insert
+     * insert，返回新插入行的 rowid（自增主键）
      *
      * @param sql
-     * @return
+     * @return 新插入行的 rowid；出错返回 -1
      */
-    public int insert(final String sql) {
+    public long insert(final String sql) {
         Log.d("flexsql", "updateSQL: " + sql);
-        return (int) db.compileStatement(sql).executeInsert();
+        return db.compileStatement(sql).executeInsert();
     }
 
     /**
